separate edit mode button into individual class.
